+++ testing.go
@@ -108,6 +108,16 @@ func (b *TestBot) Stop() {
 	b.stop()
 	err := <-b.runErr
 	if err != nil {
-		b.T.Errorf("Bot.Run() returned an error: %v")
+		b.T.Errorf("Bot.Run() returned an error: %v", err)
 	}
+}
+
+// GetOutput reads all output from b.Output and returns it as string.
+func (b *TestBot) GetOutput() string {
+	out, err := ioutil.ReadAll(b.Output)
+	if err != nil {
+		b.T.Errorf("Failed to get TestBot.Output: %v", err)
+	}
+
+	return string(out)
 }